7. Работа с датами
